♻️ refactor user profile side panel to make it reusable for public profile
