+++ frontend/src/hooks/useContributorProfilePanel/ContributorProfileSidePanel/View.tsx
@@ -1,20 +1,19 @@
 import { OwnUserProfileDetailsFragment, UserProfileFragment } from "src/__generated/graphql";
 import SidePanel from "src/components/SidePanel";
 
-import { Project } from "./ReadOnlyView/ProjectCard";
 import ReadOnlyView from "./ReadOnlyView";
 import { useEffect, useState } from "react";
 import EditView from "./EditView";
+import { UserProfile } from "./useUserProfile";
 
 type Props = {
-  profile: UserProfileFragment;
-  projects: Project[];
+  userProfile: UserProfile;
   open: boolean;
   setOpen: (value: boolean) => void;
   isOwn?: boolean;
 };
 
-export default function View({ isOwn, profile, projects, open, setOpen }: Props) {
+export default function View({ isOwn, userProfile, open, setOpen }: Props) {
   const [editMode, setEditMode] = useState(false);
   useEffect(() => {
     if (!open) setEditMode(false);
@@ -23,9 +22,12 @@ export default function View({ isOwn, profile, projects, open, setOpen }: Props)
   return (
     <SidePanel open={open} setOpen={setOpen}>
       {editMode ? (
-        <EditView profile={profile as UserProfileFragment & OwnUserProfileDetailsFragment} setEditMode={setEditMode} />
+        <EditView
+          profile={userProfile.profile as UserProfileFragment & OwnUserProfileDetailsFragment}
+          setEditMode={setEditMode}
+        />
       ) : (
-        <ReadOnlyView setOpen={setOpen} profile={profile} projects={projects} setEditMode={setEditMode} isOwn={isOwn} />
+        <ReadOnlyView setOpen={setOpen} userProfile={userProfile} setEditMode={setEditMode} isOwn={isOwn} />
       )}
     </SidePanel>
   );
