+++ frontend/src/hooks/useContributorProfilePanel/ContributorProfileSidePanel/useUserProfile.ts
@@ -1,17 +1,126 @@
 import { useQuery } from "@apollo/client";
-import { OwnUserProfileDocument, UserProfileDocument, UserProfileQuery } from "src/__generated/graphql";
+import {
+  ContributionCountFragment,
+  OwnUserProfileDocument,
+  UserProfileDocument,
+  UserProfileFragment,
+  UserProfileQuery,
+} from "src/__generated/graphql";
 import { useAuth } from "src/hooks/useAuth";
 import { contextWithCacheHeaders } from "src/utils/headers";
+import { ProfileProjectFragment } from "src/__generated/graphql";
+import { Project } from "./ReadOnlyView/ProjectCard";
+import { chain, find, range, slice, sortBy, unionBy } from "lodash";
+import onlyDustLogo from "assets/img/onlydust-logo-space.jpg";
+import { isProjectVisibleToUser } from "src/hooks/useProjectVisibility";
+import isDefined from "src/utils/isDefined";
+import { isLanguageValid } from "src/utils/languages";
+import { daysFromNow, weekNumber } from "src/utils/date";
 
-export default function useUserProfile(githubUserId: number) {
-  const { githubUserId: currentGithubUserId } = useAuth();
+export type UserProfile = {
+  profile: UserProfileFragment;
+  projects: Project[];
+  languages: string[];
+  contributionCounts: ContributionCountFragment[];
+  contributionCountVariationSinceLastWeek: number;
+};
+
+type AugmentedProfileProjectFragment = {
+  leadSince?: Date;
+} & ProfileProjectFragment;
+
+const MAX_CONTRIBUTION_COUNTS = 13;
+
+const EMPTY_DATA: ContributionCountFragment[] = range(0, MAX_CONTRIBUTION_COUNTS)
+  .map(c => daysFromNow(7 * c))
+  .map(date => ({
+    year: date.getFullYear(),
+    week: weekNumber(date),
+    paidCount: 0,
+    unpaidCount: 0,
+  }));
+
+export default function useUserProfile(githubUserId: number): UserProfile | undefined {
+  const { user: currentUser, githubUserId: currentUserGithubId } = useAuth();
 
   const { data } = useQuery<UserProfileQuery>(
-    currentGithubUserId === githubUserId ? OwnUserProfileDocument : UserProfileDocument,
+    currentUserGithubId === githubUserId ? OwnUserProfileDocument : UserProfileDocument,
     { variables: { githubUserId }, ...contextWithCacheHeaders }
   );
 
-  return {
-    userProfile: data?.userProfiles.at(0),
-  };
+  const profile = data?.userProfiles.at(0);
+
+  const projectLeaded =
+    profile?.projectsLeaded
+      .map(
+        project =>
+          ({
+            ...project.project,
+            leadSince: new Date(project.assignedAt + "Z"),
+          } as AugmentedProfileProjectFragment)
+      )
+      .filter(isDefined) || [];
+
+  const projectsContributed =
+    profile?.projectsContributed.filter(isDefined).map(
+      project =>
+        ({
+          ...project.project,
+        } as AugmentedProfileProjectFragment)
+    ) || [];
+
+  const projects: Project[] = unionBy(projectLeaded, projectsContributed, p => p.id)
+    .filter(project =>
+      isProjectVisibleToUser({
+        project,
+        user: { userId: currentUser?.id, githubUserId: currentUserGithubId },
+      })
+    )
+    .map(
+      project =>
+        ({
+          id: project.id,
+          name: project.projectDetails?.name,
+          logoUrl: project.projectDetails?.logoUrl || onlyDustLogo,
+          leadSince: project.leadSince,
+          private: project.projectDetails?.visibility === "private",
+          contributionCount: find(profile?.contributionStats, { projectId: project.id })?.totalCount,
+          lastContribution: find(profile?.contributionStats, { projectId: project.id })?.maxDate,
+          contributorCount: project.contributorsAggregate?.aggregate?.count || 0,
+          totalGranted: project.budgetsAggregate?.aggregate?.sum?.spentAmount || 0,
+        } as Project)
+    )
+    .filter(project => project.leadSince || project.contributionCount);
+
+  const languages =
+    (profile &&
+      sortBy(Object.keys(profile.languages), l => profile.languages[l])
+        .reverse()
+        .filter(isLanguageValid)) ||
+    [];
+
+  const contributionCounts =
+    (profile &&
+      chain(profile.contributionCounts)
+        .unionWith(EMPTY_DATA, (e1, e2) => e1.year === e2.year && e1.week === e2.week)
+        .sortBy(["year", "week"])
+        .reverse()
+        .take(MAX_CONTRIBUTION_COUNTS)
+        .reverse()
+        .value()) ||
+    [];
+
+  const [lastWeek, thisWeek] = slice(contributionCounts, -2);
+  const variationSinceLastWeek =
+    lastWeek && thisWeek ? thisWeek.paidCount + thisWeek.unpaidCount - (lastWeek.paidCount + lastWeek.unpaidCount) : 0;
+
+  return (
+    profile && {
+      profile,
+      projects,
+      languages,
+      contributionCounts,
+      contributionCountVariationSinceLastWeek: variationSinceLastWeek,
+    }
+  );
 }
